feat: Generation of files at 480p, 720p, and 1080p resolutions has been set up; fix: Deletion of these files and fields corrected.

+++ content/signals.py
@@ -10,24 +10,33 @@ from django_rq import enqueue
 @receiver(post_save, sender=Video)
 def video_post_save(sender, instance, created, **kwargs):
     print("VIDEO GESPEICHERT - TASK GESTARTET")
-    print("Video post save signal triggered")
     if created:
-        print("New video created:", instance.title)
         # enqueue(convert_480p, instance.video_file.path)
         # queue = django_rq.get_queue("default", autocommit=True)
         # queue.enqueue(convert_480p, instance.video_file.path)
         # django_rq.enqueue(convert_480p, instance.video_file.path)
         django_rq.enqueue(convert_480p, instance.id)
+        print("Video post 480p signal saved")
         django_rq.enqueue(convert_720p, instance.id)
-        # django_rq.enqueue(convert_1080p, instance.id)
+        print("Video post 720p signal saved")
+        django_rq.enqueue(convert_1080p, instance.id)
+        print("Video post 1080p signal saved")
 
 
 @receiver(post_delete, sender=Video)
 def auto_delete_file_on_delete(sender, instance, **kwargs):
-    """
-    Deletes file from filesystem
-    when corresponding `Video` object is deleted.
-    """
-    if instance.video_file:
-        if os.path.isfile(instance.video_file.path):
-            os.remove(instance.video_file.path)
+
+    if instance.video_file and os.path.isfile(instance.video_file.path):
+        os.remove(instance.video_file.path)
+    base_path, _ = os.path.splitext(instance.video_file.path)
+    resolutions = ["480p", "720p", "1080p"]
+
+    for res in resolutions:
+        converted_file_path = f"{base_path}_{res}.mp4"
+        if os.path.isfile(converted_file_path):
+            os.remove(converted_file_path)
+
+    for field_name in ["video_480p", "video_720p", "video_1080p"]:
+        field = getattr(instance, field_name, None)
+        if field and hasattr(field, "path") and os.path.isfile(field.path):
+            os.remove(field.path)
